0.3.1: The OCD Update
There was an indentation problem in the Level 1 file. I know it's probably irrelevant since we will be moving all the level files over to the new system very soon, but I just wanted to make sure that everything was in check in case we use this repository when we are converting the levels.