0.3.8: Fixed a tiny (but game-breaking) bug
The GUI navbar in the Testing screen wouldn't extend at all, making it impossible to enter the new play screen. The issue was somehow related to the "position: absolute" CSS attribute under the "game" ID, but I was able to fix it by putting the new "gui" span around the navbar code and setting "pointer-events: all" on the navbar itself.