0.3.25: Level select improvements
Fixed a bug which caused the level icons to stretch horizontally once they reached the triple digits. Closes #14

The player can now only navigate up to the last page that actually contains levels. For example, if the highest level is 37, the player won't be able to go to Page 4. This caps at page 4096 so that there aren't any visual glitches, so if there's a level 4097, the player won't be able to access it.

The unlocked.json file is no more! The level select screen now automatically detects which level files are present, and makes their icons blue (primary). To change the color of a level's icon, all you need to do is go into the "info" section of its JSON file and set the "icon-color" property to any of the Bootstrap colors. Closes #61

The level icons that do not have a corresponding level loaded are now not only gray, but also disabled, meaning that they cannot be clicked.